% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskOllama.R
\name{askOllama_preamble}
\alias{askOllama_preamble}
\title{ASKAI: An R Interface to Ollama API}
\description{
ASKAI (Fearless 2.0.032025) provides a streamlined interface to interact with Ollama models
through R. It offers one function with varying levels of parameter control for
different use cases.
}
\details{
Package: AskOllama
Type: Package
Version: 2.0.032025
Date: 2025-03-26
License: MIT
}
\author{
Eduardo Tamaki \href{mailto:eduardo@tamaki.ai}{eduardo@tamaki.ai}

Levente Littvay \href{mailto:levi@littvay.com}{levi@littvay.com}
}
\keyword{internal}
