% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskOllama.R
\name{list_models}
\alias{list_models}
\title{List Available Ollama Models}
\usage{
list_models()
}
\value{
A data frame containing available models and their details
}
\description{
List Available Ollama Models
}
