% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskOllama.R
\name{validate_model}
\alias{validate_model}
\title{Validate Model Selection}
\usage{
validate_model(model)
}
\arguments{
\item{model}{Character string of the model name}
}
\value{
Logical indicating if model exists
}
\description{
Validate Model Selection
}
\keyword{internal}
