% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AskOllama.R
\name{validate_numeric}
\alias{validate_numeric}
\title{Validate Numeric Parameters}
\usage{
validate_numeric(value, min, max, param_name)
}
\arguments{
\item{value}{Numeric value to validate}

\item{min}{Minimum allowed value}

\item{max}{Maximum allowed value}

\item{param_name}{Name of the parameter for error messages}
}
\value{
Logical indicating if value is valid
}
\description{
Validate Numeric Parameters
}
\keyword{internal}
